banking_accounting_rag/requirements.txt
flask
langchain
langchain-core
langchain-text-splitters
langchain-community
langchain-chroma
chromadb
ollama
python-dotenv
google-genai
openai
